AppMonitorProdutividade/MonitorProd_MVP_PostgreSQL_v1/server_postgres/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9
